+++ packages/camera/camera_android_camerax/test/android_camera_camerax_test.dart
@@ -212,6 +212,7 @@ void main() {
           AspectRatioStrategy? aspectRatioStrategy,
           ResolutionStrategy? resolutionStrategy,
           ResolutionFilter? resolutionFilter,
+          int? allowedResolutionMode,
         }) {
           final mockResolutionSelector = MockResolutionSelector();
           when(mockResolutionSelector.getAspectRatioStrategy()).thenAnswer(
@@ -219,6 +220,7 @@ void main() {
           );
           when(mockResolutionSelector.resolutionStrategy).thenReturn(resolutionStrategy);
           when(mockResolutionSelector.resolutionFilter).thenReturn(resolutionFilter);
+          when(mockResolutionSelector.allowedResolutionMode).thenReturn(allowedResolutionMode);
           return mockResolutionSelector;
         };
     PigeonOverrides.qualitySelector_from =
@@ -560,6 +562,7 @@ void main() {
             AspectRatioStrategy? aspectRatioStrategy,
             ResolutionStrategy? resolutionStrategy,
             ResolutionFilter? resolutionFilter,
+            int? allowedResolutionMode,
           }) {
             return MockResolutionSelector();
           };
@@ -759,6 +762,9 @@ void main() {
           await camera.imageAnalysis!.resolutionSelector!.resolutionStrategy!.getFallbackRule(),
           ResolutionStrategyFallbackRule.closestLowerThenHigher,
         );
+        expect(camera.preview!.resolutionSelector!.allowedResolutionMode, isNull);
+        expect(camera.imageCapture!.resolutionSelector!.allowedResolutionMode, isNull);
+        expect(camera.imageAnalysis!.resolutionSelector!.allowedResolutionMode, isNull);
       }
 
       // Test max case.
@@ -776,6 +782,19 @@ void main() {
         camera.imageAnalysis!.resolutionSelector!.resolutionStrategy,
         equals(ResolutionStrategy.highestAvailableStrategy),
       );
+      expect(
+        camera.preview!.resolutionSelector!.allowedResolutionMode,
+        ResolutionSelectorAllowedResolutionMode.preferHigherResolutionOverCaptureRate,
+      );
+      expect(
+        camera.imageCapture!.resolutionSelector!.allowedResolutionMode,
+        ResolutionSelectorAllowedResolutionMode.preferHigherResolutionOverCaptureRate,
+      );
+      expect(
+        camera.imageAnalysis!.resolutionSelector!.allowedResolutionMode,
+        ResolutionSelectorAllowedResolutionMode.preferHigherResolutionOverCaptureRate,
+      );
+      expect(camera.preview!.resolutionSelector, same(camera.imageCapture!.resolutionSelector));
 
       // Test null case.
       final int flutterSurfaceTextureId = await camera.createCamera(testCameraDescription, null);
@@ -1079,6 +1098,7 @@ void main() {
             AspectRatioStrategy? aspectRatioStrategy,
             ResolutionStrategy? resolutionStrategy,
             ResolutionFilter? resolutionFilter,
+            int? allowedResolutionMode,
           }) {
             return MockResolutionSelector();
           };
@@ -1489,6 +1509,7 @@ void main() {
           AspectRatioStrategy? aspectRatioStrategy,
           ResolutionStrategy? resolutionStrategy,
           ResolutionFilter? resolutionFilter,
+          int? allowedResolutionMode,
         }) {
           return MockResolutionSelector();
         };
@@ -1497,6 +1518,7 @@ void main() {
           AspectRatioStrategy? aspectRatioStrategy,
           ResolutionStrategy? resolutionStrategy,
           ResolutionFilter? resolutionFilter,
+          int? allowedResolutionMode,
         }) {
           return MockResolutionSelector();
         };
@@ -1829,6 +1851,7 @@ void main() {
           AspectRatioStrategy? aspectRatioStrategy,
           ResolutionStrategy? resolutionStrategy,
           ResolutionFilter? resolutionFilter,
+          int? allowedResolutionMode,
         }) => MockResolutionSelector();
     PigeonOverrides.fallbackStrategy_lowerQualityOrHigherThan = ({required VideoQuality quality}) =>
         MockFallbackStrategy();