+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.7.4+3
+
+* Updates `ResolutionPreset.max` to prefer higher resolution over capture rate
+  for CameraX `ResolutionSelector` use cases on Android.
+
 ## 0.7.4+2
 
 * Bumps cameraxVersion from 1.6.0 to 1.6.1.